fix: take the linter's remaining correctness findings
Each of these is a judgement the autofix could not make.

Production:

  * `app/api/labels.py` — a 400 raised inside an `except` now chains the
    parse error with `from e`, so the traceback does not read as a
    failure in the handler itself (B904).
  * `app/vision/local.py` — the zip over boxes, texts and scores is now
    `strict=True`. They are parallel outputs of one detection pass; if
    they ever disagree, zipping to the shortest silently drops text that
    was read, and since `f498c14` a dropped government warning is a
    rejection rather than a review. Raising is the honest outcome (B905).
  * `app/vision/local.py` — successive cut points now use
    `itertools.pairwise` (RUF007, B905).
  * `app/vision/cloud.py` — the zip over field names and gathered results
    is `strict=True`; one gather runs per field name, so the lengths are
    equal by construction and this says so (B905).
  * `app/vision/quality.py` — `int(round(x))` dropped to `round(x)`;
    one-argument `round` already returns an int (RUF046).

Tests, where the finding was that a test could not tell the difference:

  * Five `pytest.raises(Exception)` now name what they expect —
    `FrozenInstanceError` for the frozen dataclass, `ValidationError`
    matching "Instance is frozen" for the frozen pydantic models, and
    `ValidationError` for a refused setting. `Exception` also passed on
    an AttributeError from a mistyped attribute name (B017).
  * `tests/test_rule_loader_failclose.py` computed the asset's real
    digest and never used it, so it only ever showed that a wrong pin is
    refused — which a loader refusing every pin would also do. The real
    digest now drives a companion test asserting the pack loads. It
    passes, so the gap was in the test, not the loader (F841).
  * `tests/test_typescript_envelope_drift.py` tested membership against
    an empty dict, which is never true (RUF060).
  * A `match=` pattern using `.*` deliberately is now a raw string
    (RUF043).
  * `tests/fixtures/labels/make_variants.py` unpacked a size it never
    read (RUF059).

Suite: 1289 passed, 7 skipped.

+++ app/vision/quality.py
@@ -82,7 +82,7 @@ def _extract_dpi(image_bytes: bytes, dimensions: Dimensions | None) -> int | Non
     if info_dpi:
         x = float(info_dpi[0])
         if x > 0:
-            return int(round(x))
+            return round(x)
 
     exif = img.getexif()
     x_res = exif.get(_EXIF_X_RESOLUTION)
@@ -92,7 +92,7 @@ def _extract_dpi(image_bytes: bytes, dimensions: Dimensions | None) -> int | Non
         if x_val > 0:
             if unit == 3:  # cm → convert to inch
                 x_val *= 2.54
-            return int(round(x_val))
+            return round(x_val)
 
     if dimensions is not None and dimensions.dpi is not None:
         return dimensions.dpi
